# Optimize FlatVector<T>::copyRanges