@mbasmanova 2023-09-15
> for the reported benchmarks in the summary can you only report the ones that have significant change before and after the change. or at least separate those for better readability. (just in the summary)

Sure. Updated.